refactor : Running UI 요소 디렉토리 분리 #PAR-336
